editor: work with model item directly in text wrap
Instead of passing the node ID to the rich text operations wrapper, give it
a pointer to the model cursor instead and use it to modify the node.
