+++ src/editor/main.cpp
@@ -11,94 +11,7 @@
 #include "org_qml_node_edit.hpp"
 
 
-namespace org_qml {
-#define _case(__Kind)                                                     \
-    template <>                                                           \
-    struct serde<::org_qml::__Kind, ::sem::SemId<::sem::__Kind>> {        \
-        using qml_type = ::org_qml::__Kind;                               \
-        using cxx_type = ::sem::SemId<::sem::__Kind>;                     \
-        static qml_type cxx_to_qml(cxx_type value) {                      \
-            return qml_type{value};                                       \
-        }                                                                 \
-        static cxx_type qml_to_cxx(qml_type value) {                      \
-            return value.__data.as<::sem::__Kind>();                      \
-        }                                                                 \
-    };
 
-EACH_SEM_ORG_KIND(_case)
-#undef _case
-} // namespace org_qml
-
-
-class OrgNodeItem {
-  public:
-    explicit OrgNodeItem(
-        const sem::SemId<sem::Org>& data,
-        OrgNodeItem*                parent)
-        : node(data), parent(parent) {}
-
-    OrgNodeItem* at(int row) {
-        if (row < item_cache.size()) {
-            return item_cache.at(row).get();
-        } else {
-            if (row < node->subnodes.size()) {
-                auto& sub = item_cache.emplace_back(
-                    std::make_unique<OrgNodeItem>(
-                        node->subnodes.at(row), this));
-                return sub.get();
-            } else {
-                return nullptr;
-            }
-        }
-    }
-
-    int size() const { return node->subnodes.size(); }
-    int columnCount() const { return 1; }
-
-    QVariant data(int column) const {
-        if (column == 0) {
-            return QString::fromStdString(fmt1(node->getKind()));
-        } else {
-            return QVariant();
-        }
-    }
-
-    int row() const {
-        if (parent) {
-            for (int idx = 0; idx < parent->size(); ++idx) {
-                if (parent->at(idx)->node.get() == this->node.get()) {
-                    return idx;
-                }
-            }
-            return 0;
-        } else {
-            return 0;
-        }
-    }
-
-    OrgNodeItem* getparent() { return parent; }
-
-    std::vector<UPtr<OrgNodeItem>> item_cache;
-    sem::SemId<sem::Org>           node;
-    OrgNodeItem*                   parent;
-
-    QVariant getNodeRich() {
-        return QVariant::fromValue(OrgNodeTextWrapper{node});
-    }
-
-    QVariant getNodeHandle() {
-        switch (node->getKind()) {
-#define _case(__Kind)                                                     \
-    case OrgSemKind::__Kind:                                              \
-        return QVariant::fromValue(                                       \
-            org_qml::serde<org_qml::__Kind, sem::SemId<sem::__Kind>>::    \
-                cxx_to_qml(node.as<sem::__Kind>()));
-
-            EACH_SEM_ORG_KIND(_case)
-#undef _case
-        }
-    }
-};
 
 class OrgDocumentModel : public QAbstractItemModel {
   public:
@@ -122,7 +35,7 @@ class OrgDocumentModel : public QAbstractItemModel {
         const sem::SemId<sem::Org>& rootData,
         QObject*                    parent = nullptr)
         : QAbstractItemModel(parent)
-        , rootItem(std::make_unique<OrgNodeItem>(rootData, nullptr)) {}
+        , rootItem(std::make_unique<OrgNodeCursor>(rootData, nullptr)) {}
 
     ~OrgDocumentModel() override = default;
 
@@ -133,12 +46,12 @@ class OrgDocumentModel : public QAbstractItemModel {
         if (!hasIndex(row, column, parent)) {
             return QModelIndex();
         } else {
-            OrgNodeItem* itemParent = parent.isValid()
-                                        ? static_cast<OrgNodeItem*>(
-                                            parent.internalPointer())
-                                        : rootItem.get();
+            OrgNodeCursor* itemParent = parent.isValid()
+                                          ? static_cast<OrgNodeCursor*>(
+                                              parent.internalPointer())
+                                          : rootItem.get();
 
-            OrgNodeItem* childItem = itemParent->at(row);
+            OrgNodeCursor* childItem = itemParent->at(row);
             if (childItem) {
                 return createIndex(row, column, childItem);
             } else {
@@ -151,9 +64,9 @@ class OrgDocumentModel : public QAbstractItemModel {
         if (!index.isValid()) {
             return QModelIndex();
         } else {
-            OrgNodeItem* parent = static_cast<OrgNodeItem*>(
-                                      index.internalPointer())
-                                      ->getparent();
+            OrgNodeCursor* parent = static_cast<OrgNodeCursor*>(
+                                        index.internalPointer())
+                                        ->getparent();
 
             if (parent == rootItem.get()) {
                 return QModelIndex();
@@ -165,21 +78,21 @@ class OrgDocumentModel : public QAbstractItemModel {
 
     int rowCount(
         const QModelIndex& parent = QModelIndex()) const override {
-        OrgNodeItem* itemParent;
+        OrgNodeCursor* itemParent;
         if (parent.column() > 0) {
             return 0;
         } else {
-            return parent.isValid()
-                     ? static_cast<OrgNodeItem*>(parent.internalPointer())
-                           ->size()
-                     : rootItem.get()->size();
+            return parent.isValid() ? static_cast<OrgNodeCursor*>(
+                                          parent.internalPointer())
+                                          ->size()
+                                    : rootItem.get()->size();
         }
     }
 
     int columnCount(
         const QModelIndex& parent = QModelIndex()) const override {
         if (parent.isValid()) {
-            return static_cast<OrgNodeItem*>(parent.internalPointer())
+            return static_cast<OrgNodeCursor*>(parent.internalPointer())
                 ->columnCount();
         } else {
             return rootItem->columnCount();
@@ -191,12 +104,12 @@ class OrgDocumentModel : public QAbstractItemModel {
 
         if (!index.isValid()) { return QVariant(); }
 
-        OrgNodeItem* item = static_cast<OrgNodeItem*>(
+        OrgNodeCursor* item = static_cast<OrgNodeCursor*>(
             index.internalPointer());
 
         switch (role) {
             case Qt::DisplayRole: {
-                OrgNodeItem* item = static_cast<OrgNodeItem*>(
+                OrgNodeCursor* item = static_cast<OrgNodeCursor*>(
                     index.internalPointer());
                 return item->data(index.column());
             }
@@ -220,7 +133,7 @@ class OrgDocumentModel : public QAbstractItemModel {
     }
 
   private:
-    UPtr<OrgNodeItem> rootItem;
+    UPtr<OrgNodeCursor> rootItem;
 };
